Move input_symbols and output_symbols to Fst trait
